Post merge fixes
